+++ R/Interaction.R
@@ -25,6 +25,7 @@
 #' \item{grid.size: }{(`logical(1)`)\cr The number of values per feature that should be used to estimate the interaction strength.
 #' A larger grid.size means more accurate the results but longer the computation time.
 #' For each of the grid points, the partial dependence functions have to be computed, which involves marginalizing over all data points.}
+#' \item{parallel: }{`logical(1)`\cr Should the method be executed in parallel? If TRUE, requires a cluster to be registered, see ?foreach::foreach.}
 #' \item{run: }{(`logical(1)`)\cr Should the Interpretation method be run?}
 #' }
 #' 
@@ -38,13 +39,14 @@
 #' The interaction strength between two features is the proportion of the variance of the 2-dimensional partial dependence function that is not explained
 #' by the sum of the two 1-dimensional partial dependence functions.
 #' 
-#' The interaction measure takes on values between 0 (no interaction) to 1 (100% of variance of f(x) du to interaction).
+#' The interaction is measured by Friedman's H-statistic (square root of the H-squared test statistic) and takes on values between 0 (no interaction) to 1 (100% of standard deviation of f(x) du to interaction).
+#' 
 #' 
 #' @section Fields:
 #' \describe{
 #' \item{grid.size: }{(`logical(1)`)\cr The number of values per feature that should be used to estimate the interaction strength.}
 #' \item{predictor: }{(Predictor)\cr The prediction model that was analysed.}
-#' \item{results: }{(data.frame)\cr Data.frame with the interaction strength (column '.interation') per feature and - in the case of a multi-dimensional outcome - per class.}
+#' \item{results: }{(data.frame)\cr Data.frame with the interaction strength (column '.interation') per feature calculated as Friedman's H-statistic and - in the case of a multi-dimensional outcome - per class.}
 #' }
 #'  
 #' @section Methods:
@@ -74,7 +76,7 @@
 #' plot(ia) 
 #' 
 #' 
-#' # Extract the results
+#' # Extract the results 
 #' dat = ia$results
 #' head(dat)
 #' \dontrun{
@@ -93,6 +95,7 @@
 #' }
 #' }
 #' @importFrom data.table dcast
+#' @importFrom foreach %do%
 #' 
 #' @export
 NULL
@@ -103,11 +106,12 @@ Interaction = R6::R6Class("Interaction",
   public = list(
     # The fitted tree
     grid.size = NULL,
-    initialize = function(predictor, feature = NULL, grid.size = 30, run = TRUE) {
+    initialize = function(predictor, feature = NULL, grid.size = 30, run = TRUE, parallel = FALSE) {
       assert_vector(feature, len = 1, null.ok = TRUE)
       assert_number(grid.size, lower = 2)
       assert_logical(run)
-      
+      assert_logical(parallel)
+      private$parallel = parallel
       if (!is.null(feature) && is.numeric(feature)) {
         private$feature = predictor$data$feature.names[feature]
       } else {
@@ -115,16 +119,24 @@ Interaction = R6::R6Class("Interaction",
       }
       self$grid.size = min(grid.size, predictor$data$n.rows)
       super$initialize(predictor)
-      if(run) self$run()
+      if(run) self$run(predictor$batch.size)
+    },
+    run = function(batch.size) {
+      features = setdiff(private$sampler$feature.names, private$feature)
+      data.sample = private$sampler$get.x()
+      probe = self$predictor$predict(data.frame(data.sample[1,]))
+      private$multiClass = ifelse(ncol(probe) > 1, TRUE, FALSE)
+      `%mypar%` = private$get.parallel.fct(private$parallel)
+      self$results = foreach(feature = features, .combine = rbind, 
+        .export = c("self"), .packages = devtools::loaded_packages()$package, .inorder = FALSE) %mypar%   
+        private$interaction.run.single(dataSample = data.sample, 
+          feature.name = c(feature, private$feature), 
+          grid.size = self$grid.size, 
+          batch.size = batch.size, q = private$q, predictor = self$predictor)
+      private$finished = TRUE
     }
   ), 
   private = list(
-    intervene = function() {
-        intervene.interaction.multi(private$dataSample, grid.size = self$grid.size, feature.name = private$feature)
-    },
-    aggregate = function() {
-      aggregate.interaction(private$dataDesign, private$qResults, feature = private$feature)
-    }, 
     generatePlot = function(sort = TRUE, ...) {
       res = self$results
       if (sort & !private$multiClass) {
@@ -141,6 +153,50 @@ Interaction = R6::R6Class("Interaction",
         p = p + facet_wrap(".class")
       }
       p
+    },
+    interaction.run.single = function(batch.size, dataSample, feature.name, grid.size, predictor, q) {
+      assert_data_table(dataSample)
+      assert_character(feature.name, min.len = 1, max.len = 2, any.missing = FALSE)
+      assert_number(grid.size)
+      
+      grid.dat = dataSample[sample(1:nrow(dataSample), size = grid.size),]
+      dist.dat = dataSample
+      
+      res.intermediate = data.table()
+      if (length(feature.name) == 1) {
+        mg_j = MarginalGenerator$new(grid.dat, dist.dat, feature.name, cartesian = TRUE)
+        batch.size.split = floor(batch.size)/2
+        mg_noj  = MarginalGenerator$new(grid.dat, dist.dat, setdiff(colnames(dataSample),feature.name), cartesian = TRUE)
+        
+        while(!mg_j$finished) {
+          partial_j = mg_j$next.batch(batch.size.split)
+          partial_j$.type = "j"
+          partial_noj = mg_noj$next.batch(batch.size.split)
+          partial_noj$.type = "no.j"
+          grid.dat$.type = "f"
+          grid.dat$.id = 1:nrow(grid.dat)
+          res.intermediate = rbind(res.intermediate, partial_j, partial_noj, grid.dat, use.names = TRUE)
+        }
+      } else if (length(feature.name) == 2) {
+        batch.size.split = floor(batch.size/3)
+        mg_jk = MarginalGenerator$new(grid.dat, dist.dat, feature.name, cartesian = TRUE)
+        mg_j  = MarginalGenerator$new(grid.dat, dist.dat, feature.name[1], cartesian = TRUE)
+        mg_k  = MarginalGenerator$new(grid.dat, dist.dat, feature.name[2], cartesian = TRUE)
+        
+        while(!mg_j$finished) {
+          partial_jk = mg_jk$next.batch(batch.size.split)
+          partial_jk$.type = "jk"
+          partial_j = mg_j$next.batch(batch.size.split)
+          partial_j$.type = "j"
+          partial_k = mg_k$next.batch(batch.size.split)
+          partial_k$.type = "k"
+          res.intermediate = rbind(res.intermediate, partial_jk, partial_j, partial_k, use.names = TRUE)
+        }
+      }
+      predictResults = predictor$predict(data.frame(res.intermediate))
+      qResults = q(predictResults)
+      res.intermediate$.feature = paste(feature.name, collapse = ":")
+      aggregate.interaction(res.intermediate, qResults, feature.name)
     },
     feature = NULL
   )
@@ -175,7 +231,6 @@ plot.Interaction = function(x, sort = TRUE) {
 }
 
 
-
 # The test statistic as defined in:
 # Friedman, Jerome H., and Bogdan E. Popescu. "Predictive learning via rule ensembles." 
 # The Annals of Applied Statistics 2.3 (2008): 916-954.
@@ -185,12 +240,19 @@ h.test = function(f.all, f.j, f.no.j) {
   assert_numeric(f.j, any.missing = FALSE)
   assert_numeric(f.no.j, any.missing = FALSE)
   # center
-  f.all = scale(f.all, center = TRUE, scale = FALSE)
-  f.j =  scale(f.j, center = TRUE, scale = FALSE)
-  f.no.j =  scale(f.no.j, center = TRUE, scale = FALSE)
+  f.all = my.scale(f.all)
+  f.j =  my.scale(f.j)
+  f.no.j =  my.scale(f.no.j)
   # statistics
-  sum((f.all  - (f.j + f.no.j))^2) / sum(f.all^2)
+  sqrt(sum((f.all  - (f.j + f.no.j))^2) / sum(f.all^2))
 }  
+
+my.scale = function(x) {
+  x.scaled = scale(x, center = TRUE, scale = FALSE)
+  x.scaled[is.na(x.scaled)] = 0
+  x.scaled
+}
+
 
 aggregate.interaction = function(partial_dat, prediction, feature) {
   assert_data_table(partial_dat)
@@ -212,8 +274,7 @@ aggregate.interaction = function(partial_dat, prediction, feature) {
   partial_dat = partial_dat[, c(".id", ".feature", ".type", ".y.hat", ".class")]
   pd = dcast(partial_dat, .feature + .id + .class ~ .type, 
     value.var = ".y.hat", fun.aggregate = mean)
-  
-  if (length(feature) == 1) {
+  if (length(feature) == 2) {
     res = data.frame(pd[, list(.interaction = h.test(jk, j, k)), by = list(.feature, .class)])
   } else {
     res = data.frame(pd[, list(.interaction = h.test(f, j, no.j)), by = list(.feature, .class)])
@@ -224,46 +285,4 @@ aggregate.interaction = function(partial_dat, prediction, feature) {
   res
 }
 
-
-# The intervention function for the Interaction class
-# Depending on the type of interaction (1v1 or 1 vs all) creates the marginals
-intervene.interaction = function(dataSample, feature.name, grid.size) {
-  assert_data_table(dataSample)
-  assert_character(feature.name, min.len = 1, max.len = 2, any.missing = FALSE)
-  assert_number(grid.size)
-  
-  grid.dat = dataSample[sample(1:nrow(dataSample), size = grid.size),]
-  dist.dat = dataSample
-  if (length(feature.name) == 1) {
-    partial_j = generate.marginals(grid.dat, dist.dat, feature.name)
-    partial_j$.type = "j"
-    partial_noj  = generate.marginals(grid.dat, dist.dat, setdiff(colnames(dataSample),feature.name))
-    partial_noj$.type = "no.j"
-    grid.dat$.type = "f"
-    grid.dat$.id = 1:nrow(grid.dat)
-    res = rbind(partial_j, partial_noj, grid.dat, use.names = TRUE)
-  } else if (length(feature.name) == 2) {
-    partial_jk = generate.marginals(grid.dat, dist.dat, feature.name)
-    partial_jk$.type = "jk"
-    partial_j  = generate.marginals(grid.dat, dist.dat, feature.name[1])
-    partial_j$.type = "j"
-    partial_k  = generate.marginals(grid.dat, dist.dat, feature.name[2])
-    partial_k$.type = "k"
-    res = rbind(partial_jk, partial_j, partial_k, use.names = TRUE)
-  }
-  res$.feature = paste(feature.name, collapse = ":")
-  res
-}
-
-
-intervene.interaction.multi = function(dataSample, grid.size, feature.name = NULL) {
-  features = setdiff(colnames(dataSample), feature.name)
-  res = lapply(features, function (feature) {
-    dt = intervene.interaction(dataSample = dataSample, 
-      feature.name = c(feature, feature.name), 
-      grid.size = grid.size)
-    dt
-  })
-  rbindlist(res, use.names = TRUE)
-}