+++ R/InterpretationMethod.R
@@ -8,9 +8,9 @@ InterpretationMethod = R6::R6Class("InterpretationMethod",
       private$plotData = private$generatePlot(...)
       if (!is.null(private$plotData)) {
         return(private$plotData)
-        } else {
-          warning("call run() first!")
-        }
+      } else {
+        warning("call run() first!")
+      }
     },
     initialize = function(predictor) {
       checkmate::assert_class(predictor, "Predictor")
@@ -37,9 +37,7 @@ InterpretationMethod = R6::R6Class("InterpretationMethod",
         private$dataSample = private$getData()
         private$dataDesign = private$intervene()
         # EXECUTE experiment
-        private$predictResults = self$predictor$predict(data.frame(private$dataDesign))
-        private$multiClass = ifelse(ncol(private$predictResults) > 1, TRUE, FALSE)
-        private$qResults = private$q(private$predictResults)
+        private$qResults = private$run.prediction(private$dataDesign)
         # AGGREGATE measurements
         self$results = data.frame(private$aggregate())
         private$finished = TRUE
@@ -47,6 +45,7 @@ InterpretationMethod = R6::R6Class("InterpretationMethod",
     }
   ),
   private = list(
+    parallel = FALSE,
     # The sampling object for sampling from X
     sampler = NULL,
     # Wrapper for sampler
@@ -77,6 +76,18 @@ InterpretationMethod = R6::R6Class("InterpretationMethod",
       self$results = NULL
       private$finished = FALSE
     }, 
+    run.prediction = function(dataDesign) {
+      private$predictResults = self$predictor$predict(data.frame(dataDesign))
+      private$multiClass = ifelse(ncol(private$predictResults) > 1, TRUE, FALSE)
+      private$q(private$predictResults)
+    },
+    get.parallel.fct = function(parallel = FALSE) {
+      if(parallel) {
+        foreach::`%dopar%`
+      } else {
+        foreach::`%do%`
+      }
+    },
     # The data need for plotting of results
     plotData = NULL,
     # Function to generate the plot